before create new model of filters

+++ database/migrations/2020_05_01_140204_create_article_tag_table.php
@@ -15,7 +15,7 @@ class CreateArticleTagTable extends Migration
     {
         Schema::create('article_tag', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('article_id')->unsigned();
+            $table->bigInteger('article_id')->unsigned();
             $table->foreign('article_id')->references('id')->on('articles');
             $table->integer('tag_id')->unsigned();
             $table->foreign('tag_id')->references('id')->on('tags');
